refactor: restructure modules and clean up tool registries
