+++ src/formatters/shfmt.ts
@@ -1,11 +1,11 @@
 import { installer } from '@/installer'
 import { exec } from '@/exec'
-import { buildDockerRunArgs } from '@/utils'
+import { buildDockerRunArgs } from '@/builders'
 import type { Runner } from '@/types'
 
 const toolName = 'shfmt'
 
-export const runner: Runner = async (version, args, paths) => {
+export const runner: Runner = async ({ version, args, paths }) => {
     const tag = version === 'latest' ? 'v3' : `v${version}`
 
     await installer(toolName, tag)
