+++ src/linters/alex.ts
@@ -4,7 +4,7 @@ import type { Runner } from '@/types'
 
 const toolName = 'alex'
 
-export const runner: Runner = async (version, args, paths) => {
+export const runner: Runner = async ({ version, args, paths }) => {
     await installer(toolName, version)
 
     return await exec(toolName, [...args, '--', ...paths])
