+++ src/linters/cargo-clippy.ts
@@ -5,7 +5,7 @@ import type { Runner } from '@/types'
 const toolName = 'cargo-clippy'
 
 // NOTE: This linter doesn't support file path input.
-export const runner: Runner = async (version, args) => {
+export const runner: Runner = async ({ version, args }) => {
     await installer(toolName, version === 'latest' ? 'stable' : version)
 
     return await exec(toolName, args)
