+++ src/linters/cargo-msrv.ts
@@ -4,7 +4,7 @@ import type { Runner } from '@/types'
 
 const toolName = 'cargo-msrv'
 
-export const runner: Runner = async (version, args) => {
+export const runner: Runner = async ({ version, args }) => {
     await installer(toolName, version)
 
     return await exec(toolName, ['verify', ...args])
