+++ src/linters/shellcheck.ts
@@ -1,11 +1,11 @@
 import { installer } from '@/installer'
 import { exec } from '@/exec'
-import { buildDockerRunArgs } from '@/utils'
+import { buildDockerRunArgs } from '@/builders'
 import type { Runner } from '@/types'
 
 const toolName = 'shellcheck'
 
-export const runner: Runner = async (version, args, paths) => {
+export const runner: Runner = async ({ version, args, paths }) => {
     const tag = version === 'latest' ? 'stable' : `v${version}`
 
     await installer(toolName, tag)
