+++ src/linters/trivy.ts
@@ -1,11 +1,12 @@
 import { installer } from '@/installer'
 import { exec } from '@/exec'
-import { fileExists, buildDockerRunArgs } from '@/utils'
+import { fileExists } from '@/utils'
+import { buildDockerRunArgs } from '@/builders'
 import type { Runner } from '@/types'
 
 const toolName = 'trivy'
 
-export const runner: Runner = async (version, args) => {
+export const runner: Runner = async ({ version, args }) => {
     const hasPackageJson = await fileExists('package.json')
 
     await installer(toolName, version, { hasPackageJson })
