+++ src/utils.ts
@@ -1,22 +1,4 @@
-import { cwd } from 'node:process'
 import { access, constants } from 'node:fs/promises'
-
-export function buildDockerRunArgs(imageName: string): string[] {
-    return ['run', '--rm', '-v', `${cwd()}:/mnt`, '-w', '/mnt', imageName]
-}
-
-export function unorderedList(items: string[]): string {
-    return items.map((item) => `- ${item}`).join('\n')
-}
-
-// https://github.com/microsoft/TypeScript/issues/14520
-export function isIncluded<T extends string>(value: string, values: T[]): value is T {
-    return values.includes(value as T)
-}
-
-export function getKeys<T extends object>(obj: T): (keyof T)[] {
-    return Object.keys(obj) as (keyof T)[]
-}
 
 export async function fileExists(...paths: string[]): Promise<boolean> {
     try {
@@ -29,4 +11,13 @@ export async function fileExists(...paths: string[]): Promise<boolean> {
     } catch {
         return false
     }
+}
+
+export function getKeys<T extends object>(obj: T): (keyof T)[] {
+    return Object.keys(obj) as (keyof T)[]
+}
+
+// https://github.com/microsoft/TypeScript/issues/14520
+export function isIncluded<T extends string>(value: string, values: T[]): value is T {
+    return values.includes(value as T)
 }
